requirements.txt
black==24.4.2
click==8.1.7
colorama==0.4.6
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.2.1
python-dotenv==1.0.1
